% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{muscle_models}
\alias{muscle_models}
\title{Reference Models for Body Composition}
\format{
\subsection{\code{muscle_models}}{

A named list of skeletal muscle models. Naming scheme: (metric)\emph{(sex)}(vertebral level)
}
}
\source{
\url{https://bodycomp-metrics.mgh.harvard.edu}
}
\usage{
muscle_models
}
\description{
LMSP models for age- and sex-specific reference values for CT-assessed skeletal muscle.
}
\keyword{datasets}
